+++ .github/workflows/beta-deployment.yml
@@ -33,11 +33,3 @@ jobs:
       setupsigning: true
       fastlanelane: beta
     secrets: inherit
-  deployfirebase:
-    name: Deploy Firebase Project
-    needs: iosapptestflightdeployment
-    uses: StanfordBDHG/.github/.github/workflows/firebase-deploy.yml@v2
-    permissions:
-      contents: read
-    secrets:
-      GOOGLE_APPLICATION_CREDENTIALS_BASE64: ${{ secrets.GOOGLE_APPLICATION_CREDENTIALS_BASE64 }}
