+++ ENGAGEHF/Resources/Localizable.xcstrings
@@ -88,6 +88,16 @@
         }
       }
     },
+    "accountDeletionError" : {
+      "localizations" : {
+        "en" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "Account deletion within the app is not currently supported. Please contact your study coordinator."
+          }
+        }
+      }
+    },
     "Add Mock" : {
 
     },
@@ -169,51 +179,6 @@
     },
     "Graph Selection" : {
 
-    },
-    "HEALTHKIT_PERMISSIONS_BUTTON" : {
-      "extractionState" : "stale",
-      "localizations" : {
-        "en" : {
-          "stringUnit" : {
-            "state" : "translated",
-            "value" : "Grant Access"
-          }
-        }
-      }
-    },
-    "HEALTHKIT_PERMISSIONS_DESCRIPTION" : {
-      "extractionState" : "stale",
-      "localizations" : {
-        "en" : {
-          "stringUnit" : {
-            "state" : "translated",
-            "value" : "This onboarding step allows you to customize the onboarding flow to explain how the application uses the HealhtKit data and allows a user to cusomize the selection."
-          }
-        }
-      }
-    },
-    "HEALTHKIT_PERMISSIONS_SUBTITLE" : {
-      "extractionState" : "stale",
-      "localizations" : {
-        "en" : {
-          "stringUnit" : {
-            "state" : "translated",
-            "value" : "Spezi can access data from HealthKit using the HealthKitDataSource module."
-          }
-        }
-      }
-    },
-    "HEALTHKIT_PERMISSIONS_TITLE" : {
-      "comment" : "MARK: HealthKit",
-      "extractionState" : "stale",
-      "localizations" : {
-        "en" : {
-          "stringUnit" : {
-            "state" : "translated",
-            "value" : "HealthKit Access"
-          }
-        }
-      }
     },
     "Heart Health" : {
 
@@ -358,6 +323,16 @@
           "stringUnit" : {
             "state" : "translated",
             "value" : "Interesting Modules"
+          }
+        }
+      }
+    },
+    "invalidHKSample" : {
+      "localizations" : {
+        "en" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "Sample type not supported."
           }
         }
       }
@@ -581,6 +556,16 @@
     },
     "User not authenticated. Please sign in an try again." : {
       "comment" : "User Not Authenticated"
+    },
+    "userNotSignedIn" : {
+      "localizations" : {
+        "en" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "User not signed in."
+          }
+        }
+      }
     },
     "vitalsBloodPressure" : {
       "comment" : "For Heart Health view",
